Rename form schema files.
