+++ app/javascript/components/flavor-form/flavor-form.schema.js
